@@ -8,16 +8,16 @@
   },
   "dependencies": {
     "lodash": "^4.17.21",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "slate": "^0.76.1",
     "slate-history": "^0.66.0",
     "slate-react": "^0.76.1"
   },
   "devDependencies": {
     "@types/lodash": "^4.14.170",
-    "@types/react": "^17.0.11",
-    "@types/react-dom": "^17.0.8",
+    "@types/react": "^18.0.1",
+    "@types/react-dom": "18.0.0",
     "@vitejs/plugin-react-refresh": "^1.3.1",
     "typescript": "^4.3.4",
     "vite": "^2.3.8"
